backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
supabase==2.3.4
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# MQTT
paho-mqtt==2.0.0

# AI Integration
anthropic==0.18.1
openai==1.10.0

# WebSockets & Realtime
python-socketio==5.11.1
websockets==12.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (Compatible with supabase)
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0